Unbreak futures-cpupool and FuturesUnordered
